+++ src/components/empty-states/index.ts
@@ -3,4 +3,5 @@ export { TodayEmptyState } from "./today-empty-state";
 export { UpcomingEmptyState } from "./upcoming-empty-state";
 export { CompletedEmptyState } from "./completed-empty-state";
 export { TrashEmptyState } from "./trash-empty-state";
-export type { EmptyStateTaskInput } from "./task-add-dialog";
+export type { EmptyStateTaskInput } from "@/components/task";
+
